python/ms-prontuario-animal-main/ms-prontuario-animal/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
requests
python-multipart
alembic